rename rainbowup to rainbow, enviroment to rainbow_cli for npm
