+++ rainbow_cli/.eslintignore
