+++ rainbow_cli/.eslintrc
