+++ rainbow_cli/.gitignore
