+++ rainbow_cli/Dockerfile
