+++ rainbow_cli/commands/clean.js
@@ -10,12 +10,12 @@ class CleanCommand {
                 process.exit(1);
             }
             try {
-                execSync('python3 ~/.rainbowup/nearup/main.py stop');
+                execSync('python3 ~/.rainbow/nearup/main.py stop');
             } catch (err) {
                 console.log(`Error stopping nearup ${err}`);
             }
-            console.log('Cleaning ~/.rainbowup , ~/.nearup , and ~/.near directories...');
-            execSync('rm -rf ~/.rainbowup && rm -rf ~/.nearup && rm -rf ~/.near');
+            console.log('Cleaning ~/.rainbow directory');
+            execSync('rm -rf ~/.rainbow');
             console.log('Cleaning done...');
             process.exit(0);
         });
