+++ rainbow_cli/commands/eth-dump.js
