+++ rainbow_cli/commands/init-eth-ed25519.js
