+++ rainbow_cli/commands/init-eth-erc20.js
