+++ rainbow_cli/commands/init-eth-locker.js
