+++ rainbow_cli/commands/init-near-contracts.js
