+++ rainbow_cli/commands/init-near-fun-token.js
