+++ rainbow_cli/commands/init-near2eth-client.js
