+++ rainbow_cli/commands/init-near2eth-prover.js
