+++ rainbow_cli/commands/prepare.js
