+++ rainbow_cli/commands/start/helpers.js
