+++ rainbow_cli/commands/stop/near.js
@@ -3,7 +3,7 @@ const { execSync } = require('child_process');
 class StopLocalNearNodeCommand {
     static execute () {
         console.log('Stopping local near node...');
-        const command = 'python3 ~/.rainbowup/nearup/main.py stop';
+        const command = 'python3 ~/.rainbow/nearup/main.py stop';
         try {
             execSync(command);
         } catch (err) {
