+++ rainbow_cli/commands/stop/process.js
