+++ rainbow_cli/commands/transfer-eth-erc20-to-near.js
