+++ rainbow_cli/index.js
@@ -3,13 +3,13 @@ const { program } = require('commander');
 
 const { CleanCommand } = require('./commands/clean');
 const { PrepareCommand } = require('./commands/prepare');
-const { StartEthRelayCommand } = require('./commands/start/eth-relay.js');
-const { StartNearRelayCommand } = require('./commands/start/near-relay.js');
-const { StartNearWatchdogCommand } = require('./commands/start/near-watchdog.js');
-const { StartGanacheNodeCommand } = require('./commands/start/ganache.js');
-const { StartLocalNearNodeCommand } = require('./commands/start/near.js');
-const { StopLocalNearNodeCommand } = require('./commands/stop/near.js');
-const { StopManagedProcessCommand } = require('./commands/stop/process.js');
+const { StartEthRelayCommand } = require('./commands/start/eth-relay.js.js');
+const { StartNearRelayCommand } = require('./commands/start/near-relay.js.js');
+const { StartNearWatchdogCommand } = require('./commands/start/near-watchdog.js.js');
+const { StartGanacheNodeCommand } = require('./commands/start/ganache.js.js');
+const { StartLocalNearNodeCommand } = require('./commands/start/near.js.js');
+const { StopLocalNearNodeCommand } = require('./commands/stop/near.js.js');
+const { StopManagedProcessCommand } = require('./commands/stop/process.js.js');
 const { TransferETHERC20ToNear } = require('./commands/transfer-eth-erc20-to-near');
 const { TransferEthERC20FromNear } = require('./commands/transfer-eth-erc20-from-near');
 const { InitETHLocker } = require('./commands/init-eth-locker');
