+++ rainbow_cli/lib/borsh/nearProof.js
