+++ rainbow_cli/lib/eth-proof-extractor/index.js
