+++ rainbow_cli/lib/eth-prover-contract/index.js
