+++ rainbow_cli/lib/eth2near-client-contract/dag_merkle_roots.json
