+++ rainbow_cli/lib/near-helpers.js
