+++ rainbow_cli/lib/near-mintable-token.js
