+++ rainbow_cli/lib/near2eth-watchdog.js
