+++ rainbow_cli/scripts/start_ganache.sh
